refactor: Home page
    - created layout per section
    - move all the constants to global constants
    - add types for the directory
