+++ components/layouts/index.ts
@@ -1,0 +1,21 @@
+import AboutUs from "./AboutUs";
+import Collections from "./Collections";
+import Community from "./Community";
+import Directories from "./Directories";
+import Header from "./Header";
+import Metrics from "./Metrics";
+import PartnersSection from "./PartnersSection";
+import Pillars from "./Pillars";
+import SubCommunities from "./SubCommunities";
+
+export {
+  AboutUs,
+  Collections,
+  Community,
+  Directories,
+  Header,
+  Metrics,
+  PartnersSection,
+  Pillars,
+  SubCommunities,
+};
